pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "agentbasis"
version = "0.1.0"
description = "Management & Observability SDK for AI Agents"
readme = "README.md"
requires-python = ">=3.8"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    "requests>=2.25.0",
    "pydantic>=2.0.0",
    "opentelemetry-api>=1.20.0",
    "opentelemetry-sdk>=1.20.0",
    "opentelemetry-exporter-otlp-proto-http>=1.20.0"
]

[project.urls]
"Homepage" = "https://agentbasis.co"
"Bug Tracker" = "https://github.com/AgentBasis-AI/agentbasis-python-sdk/issues"
